Testing, Refactoring and Project Restructure
# Testing, Refactoring and Project Restructure
### pytest
 - Installed pytest framework to start testing the code and ensure quality and robustness.
### Project Restructure
 - Moved all *_service* classes to a dedicated *services* folder.
 - Added an `__init__.py__` file to ensure it behaves as a python package.
### Testing
 - Created a dedicated *tests* folder for testing.
 - Added `__init.py__` file to make sure it also behaves as a package.
 - Created a `test_gtts_service.py` file to test the *gTTS_Service* class.
 - Created 2 unit tests to check the constructor `gTTS_Service()`, see TC-001 and TC-002.
### Refactoring
 - In the class *gTTS_Service*, removed ability to determine the default directory to save files to, the value is now hard coded to prevent user abuse.
