+++ backend/services/script_writer_service.py
