+++ backend/services/service_manager.py
@@ -1,6 +1,6 @@
-from research_service import TavilyResearchService
-from script_writer_service import GeminiWriterService
-from tts_service import gTTS_Service
+from .research_service import TavilyResearchService
+from .script_writer_service import GeminiWriterService
+from .tts_service import gTTS_Service
 from datetime import datetime
 
 class ServiceManager():
